refactor: Migrate MyAppointmentsScreen to MVI
This commit refactors the `MyAppointmentsScreen` by:
- Deleting the old `MyAppointmentsScreen.kt` file.
- Renaming `MyAppointmentsContent.kt` to `MyAppointmentsScreen.kt` and moving it to the `appointments` subpackage.
- Updating the new `MyAppointmentsScreen` to adopt the Model-View-Intent (MVI) architecture.

**Key changes in the new `MyAppointmentsScreen`:**
- Integrates `AppointmentListViewModel` (using Koin for dependency injection).
- Manages UI state (`AppointmentListState`) and handles UI events (`AppointmentListEvent`).
- Uses `LaunchedUiEffectHandler` to manage side effects (`AppointmentListEffect`) like navigation and showing messages.
- Displays success and error messages using `SuccessesDialog` and `DialogWithIcon` respectively.
- Renames string resources to follow a more consistent naming convention (e.g., `R.string.my_appointments` to `R.string.features_appointments_my_appointments`).
- Makes `UpcomingAppointmentsContent` and `OtherAppointmentsContent` private as they are only used within this screen.
- Updates preview function `MyAppointmentContentPreview` to be private.

**Affected files:**
- **features/appointments/src/patient/kotlin/eg/edu/cu/csds/icare/appointment/screen/MyAppointmentsScreen.kt:** (deleted)
- **features/appointments/src/patient/kotlin/eg/edu/cu/csds/icare/appointment/screen/appointments/MyAppointmentsScreen.kt:** (renamed from `features/appointments/src/patient/kotlin/eg/edu/cu/csds/icare/appointment/screen/MyAppointmentsContent.kt`)
    - Updated package declaration.
    - Implemented MVI pattern with `AppointmentListViewModel`.
    - Added `LaunchedUiEffectHandler` for side effects.
    - Updated string resource keys.
    - Changed visibility of internal composables and preview function.
